separating TWAdvancedSearchPanel
